requirements.txt

#git+https://github.com/nvie/pip-tools.git@future#egg=pip-tools
BeautifulSoup4==4.3.2
FFVideo==0.0.13
PIL==1.1.7
PyYAML==3.10
Pympler==0.3.1
avro==1.7.6
boto==2.32.1
coverage==3.7
dateutils==0.6.6
flickrapi==1.4.2
futures==2.1.5
google-api-python-client==1.1
happybase==0.9
httpagentparser==1.6.0
imagehash==0.3
impyla==0.8.1
ipdb==0.8
matplotlib==1.3.1
mock==1.0.1
moto==0.2.9
mrjob==0.4.1
nose-cov==1.6
pandas==0.15.2
paramiko==1.14.0
poster==0.8.1
psutil==1.2.1
pyfakefs==2.4
pygeoip==0.2.7
pyhs2==0.6.0
pyleargist==2.0.5
pylint==1.1.0
redis==2.10.3
scipy==0.9.0
shortuuid==0.3
simplejson==2.3.2
svmlight==0.4
tornado==4.2.1
thrift==0.9.1
tornado==4.2.1
winpdb==1.4.8
voluptuous==0.8.7
scikit-learn==0.16.1
psycopg2==2.6.1
queries==1.7.5
PyJWT==1.4.0
momoko==2.2.2
xlwt==1.0.0
dill==0.2.4
passlib==1.6.5
youtube_dl
